Avoid copying in scanNumber
Summary:
Now that we've migrated to `fastStrToDouble`, this has unlocked direct
support for UTF16 input strings. We no longer need to be copying the
contents of the number to a local vector. Instead, we can just pass a
view over the string inputs directly to `fastStrToDouble`.

Since we also support UTF8 via the streaming API, we also have to
support that case by calling the pair of beginCapture / endCapture.

Reviewed By: avp

Differential Revision: D86346460

fbshipit-source-id: ee9cf31abe8518b18c47669b65d3f0ad9918a395