Add polyfill for the Iterator object
Summary:
Adapted from the iterator helper polyfill in
https://github.com/es-shims/iterator-helpers

This adds the polyfill for the global `Iterator` object and its
constructor.  `Iterator.prototype` methods will be added in future
changes.

Note that the original polyfill had many dependencies. The dependencies
are taken and flattened in `05-IteratorHelperDependencies.js`.
 Some dependencies are purposely taken out since Hermes already
supports it and does not require a polyfill for it.

The actual Iterator object is added in `06-IteratorObject.js`

Reviewed By: avp

Differential Revision: D84199898

fbshipit-source-id: 7e6ed0fc578cb6e368f240a28cafbcfc2bc33613