Implement serialization with transfer
Summary:
Implements the `serializeWithTransfer` and `deserializeWithTransfer` of
the `ISerialization` interface, allowing `ArrayBuffer`s to be
transferred between runtime instance.

When an ArrayBuffer is transferred, it is detached from the current
runtime, but the underlying storage is not freed. The ownership of the
underlying buffer is transferred to the SerializedValue object.

For internal buffers, it is enough to store the pointer the underlying
data, and set the internal data block directly when transferring. For
external buffers, we store the data pointer and the context from the
NativeState that manages the data block. When transferring, the
context is used for the new NativeState.

If the ArrayBuffer is never deserialized successfully in the new
runtime, it is the responsibility of the SerializedObject to destroy
the data in its destructor by freeing the data pointer for internal data
blocks. Nothing needs to be done for external data blocks because the
ownership will be released when the stored shared pointer gets
destroyed.

Reviewed By: tmikov

Differential Revision: D80988951

fbshipit-source-id: 24458cd169edd9ac0f8928d3d76d64dd8ee59fae