Use shared pointer to set JS Array Buffer external blocks
Summary:
When setting an external data block, take in shared_pointer<void> that
manages the lifetime of the underlying data.

Then, create a NativeState that stores the shared pointer as the
context. GC will release the shared pointer when it finalizes the JS
ArrayBuffer, triggering the deleter if necessary.

This allows us to transfer the ownership of the shared pointer out to
the serialized object when serializing JS Array Buffers later in the
stack.

Reviewed By: tmikov, lavenzg

Differential Revision: D80120893

fbshipit-source-id: 66ad592fa5f7725c4926bc52bb13e1b6a70b1971