Implement ISerialization interface
Summary:
Implements the `ISerialization` interface for Hermes runtime, following
the structured clone algorithm.

# Overview

Each JS value is recorded as a Record with the format `(type tag, object
id, content)`

The type tag will indicate the type of the JS value this record
encapsulates, the ID is for references used by other Records, and the
content will contain everything necessary to deserialize and reconstruct
the value. There will also be a "Reference" type, which will simply
refer to an object id, if the same object is referred to in multiple
places.

Serialization will be written to the `SerializedValue` object introduced
in this diff. At a high-level, there are 3 components:

* `offsets` buffer: Used to map an ID to the offset at which some value
is encoded
* `content` buffer: Used to store all Records, including any nested
records
* `strings` buffer: Used to store strings in the format `(string ID,
isAscii, length, data)`

# Strings

All strings will be written in the `strings` buffer with the format
`(string ID, isAscii, length, data)`. For each string encoding, the
`offsets` buffer will have a corresponding entry that identifies the
offset at which the encoding starts.

String Primitives and String Objects will simply refer to the string ID
for the content using the format `(type tag, string id)`

For places where string contents are expected directly (e.g. Object
property keys), it will simply refer to the string id.

# Cache

During serialization/deserialization, we add an occupied entry to a
`ManagedChunkList`. All occupied entries in the list will be marked
during GC's root marking. At the end of the
serialization/deserialization process, we iterate through all entries
that were added, and mark them as free.

For serialization, we can use this to track values that has been
serialized to the serialization ID to prevent serializing objects
multiple time. Similarly, for deserialization, we maintain a mapping of
serialization ID to its deserialized form.

Reviewed By: tmikov

Differential Revision: D76547680

fbshipit-source-id: 16c672050076080592ea5db3e1c1896a92b8bd3c