Allow Internal Properties to be added/updated while Object is sealed/frozen
Summary:
Currently, if the an Object is sealed, updating an internal property
will fail. Similarly of an Object is frozen, both adding and updating an
internal property will fail.

Internal properties are often an implementation detail, and writes
should be able to succeed regardless of the frozen/seal state. This
change will add `InternalForce` to force these property writes to
succeed. Note some internal properties are added during object creation,
and never updated. It is impossible for the writes to happen when Object
is frozen/sealed in this case, so no flag is needed.

NativeStates are a special case, we are treating it as a "property" and
will remain to be unmodifiable during frozen/sealed state.

Reviewed By: avp

Differential Revision: D80855592

fbshipit-source-id: fabfbc18212d733b1e9fa36291dd9de7ac9f428c