Add `createWithInternalDataBlock` and `getExternalDataContext` API to JSArrayBuffer
Summary:
Add a few new API for `ArrayBuffer` serialization and deserialization

Specially:
*  `getExternalDataContext`: for external buffers, get the Native State
context of the external data block. This allows the data lifetime to be
shared with the Serialized object via shared ptr
* `createWithInternalDataBlock`: creates `JSArrayBuffer` with a data
block. If there is an existing data attached, detach from it and replace
with the provided buffer.

Used in next serialization diff to deserialize JSArrayBuffers with
internal data blocks.

Reviewed By: fbmal7, dannysu

Differential Revision: D79815362

fbshipit-source-id: 9bdec5a0c365ae305f37c4f216b80396704fbf67