Allow ArrayBuffer to detach without cleaning up storage
Summary:
Adds a API `JSArrayBuffer::ejectBufferUnsafe` to detach the underlying
storage from the ArrayBuffer, but not clean up the underlying storage.

For internal buffers, do not free the memory, but untrack it from GC.

For external buffers, set the finalizers to do nothing.

Used in the next diff for serialization, where buffers can be detached,
but the storage ownership needs to be transferred.

Reviewed By: avp

Differential Revision: D79765779

fbshipit-source-id: ef468eb7482398d53c73949c8070cf09f46a358c