Parser: Error on extra comma in async arrow function params
Summary:
The rest parameter of an async arrow function must not be followed by a
`,`, but spreads of call expressions may be followed by `,`.

To handle this, we need to parse a `CoverTrailingCommaNode` and delete
it when running `SemanticResolver` on the resulting AST.
There's no other way to signal to the `reparse` that there was a
trailing comma in the original call expression that's being reparsed
into an async arrow function.

Reviewed By: tmikov

Differential Revision: D88689844

fbshipit-source-id: 7c131ebe4903e7563cf12c11ee8641980ef3511f