VM: Delete unnecessary MustExist asserts for Proxy object
Summary:
The `MustExist` flag doesn't really make sense to enforce for Proxy
objects.

If you have the code:

```
"use strict";
var p = new Proxy({},{});
globalThis.__proto__ = p;
result = 0;
```

It'll cause an assertion failure because the parent of the global object
is a proxy and the property write function tries to call the Proxy trap,
but that's not allowed because the MustExist flag is set.

So delete the assert because ignoring the flag is reasonable for Proxy.

Reviewed By: tmikov

Differential Revision: D88896922

fbshipit-source-id: 0ef5ebdc91f039250685687929d144a054edde3d