VM: Throw when setting the parent of the global object
Summary:
The parent of the global object is actually host-defined.
Certain engines like JSC simply make it immutable.
That's what this diff does as well.
Now it's like `Object.prototype` (which also has an immutable
prototype, but that's required by the spec).

Other engines such as V8 don't throw when modifying the parent of the
global object, but they don't call the Proxy trap for setting properties
either. They just set the property on the global object itself.

Reviewed By: tmikov

Differential Revision: D88428865

fbshipit-source-id: e92e83795adeb3ecc9a57f845f3d64f5f4fc3f52