Fix ID_Continue detection for unicode characters
Summary:
`ID_Continue` was being detected with a check that was mostly correct,
but unlike `ID_Start` it didn't take all the possible unicode codepoints
into account.

Fix this by adding a few extra numbers to `UnicodeData.inc` (making sure
to use the same trick as `ID_Start` to avoid bloating the file).

Reviewed By: fbmal7, tmikov

Differential Revision: D88788365

fbshipit-source-id: 723aec03843e05a5e7fc69ceb6336d5aaa91a5dd