Parser: Populate identifiers from Keywords
Summary:
Many of these already exist in `Keywords`, and it's confusing
to have two sources of truth to populate every time.

The second step to this would have been to replace all these member
fields of `JSParserImpl` with direct access to `kw_`, but I don't want
to cause tons of source control noise in this diff.

In the future, we can just add to `Keywords` which is more convenient
and forego the need to add member variables.

Reviewed By: tmikov

Differential Revision: D89093914

fbshipit-source-id: 4d650f7d0743062085378c7b5198491f017245a0