Remove unnecessary `llvh::dyn_cast` usages
Summary:
`llvh::dyn_cast` and `llvh::dyn_cast_or_null` is used in
SynthTraceParser, but we already know the format of the Synth Trace
records. This means that we know the type of each field, so we can just
use `llvh::cast`, or `llvh::cast_or_null`.

For fields that must exist on the Record, use `llvh::cast`. Otherwise,
use `llvh::cast_or_null` if the field is optional.

Reviewed By: lavenzg

Differential Revision: D88680402

fbshipit-source-id: c7c8ac1b3552eeb62b1195fbf6a7a61119b1c3e2