Adding serialization/deserialization support in SynthTraces
Summary:
Adds Synth Trace support for `ISerialization` interface.

This is the first time an optional API is added to Synth Trace.
`TracingRuntime` will inherit the optional interface and provide its
implementation to create the appropriate trace records and call the
underlying runtime's version. When `castInterface` is called, if the
underlying Runtime supports the API, Tracing Runtime will return a
pointer to its implementation. Otherwise, return `nullptr`.

Serialization Records: We simply record the ID of the value being
serialized.

Deserialization Records: We need to store the Serialized object in the
`trace.json` file so we can replay the deserialization. We also
deserialize, and create a `ReturnToNative` record to record the newly
created `Value`.

Transfer is purposely not supported since we cannot meaningfully
transfer objects in synth traces. This is consistent with the HTML spec
(`StructuredSerializeForStorage` does not deal with transfers).

Reviewed By: lavenzg

Differential Revision: D87683188

fbshipit-source-id: 355494108e08f219bd1168d1f5d6be879a739d75