Emit function declarations after `this` initialization for legacy classes
Summary:
`emitFunctionPrologue` can potentially generate IR for a `super()` call,
if that appears as the default value for a parameter. When setting up a
class constructor, the state needed to support `super()` calls is not
set up until after `emitFunctionPrologue` runs. So we hit an assert. The
fix is: for class constructors, pass in `DoEmitDeclarations::No` to
`emitFunctionPrologue`. Factor out the code that's gated behind
`DoEmitDeclarations` into it's own function. Then once the `super()`
state is seeded, manually call out to this refactored code from
`genBasicFunction`.

Reviewed By: avp

Differential Revision: D86111330

fbshipit-source-id: de26cbf51ec7d849a3f825cfa8ea78679f5397f1