Fix flaky weakmap.js, weakset.js and weakmap-key-in-value.js‎ test
Summary:
When calling `HermesInternal.getWeakSize()`, now we may trigger a GC
when handle sanitizer is ON. It may collect the entry of key `b`
because `b` is never used after the call in the local function. So the
test may fail.

Add a reference to `b` after the `getWeakSize()` call to keep it alive.

To capture such errors easier in the future, I also set the sanitize
rate to 0.1 for three weakmap/weakset tests.

Reviewed By: avp

Differential Revision: D89488875

fbshipit-source-id: 88b7c50150437debd5bd3e08debea3b44bfacd87