Fix wrong int type used in incrementAllocatedBytes()
Summary:
We use int32_t for parameter of `incrementAllocatedBytes()`. Since
HadesGC now supports large allocation, the size of a single allocation
could be larger than 1 << 31, which will overflow int32_t. This is
undefined behavior. In practice, on ARM64 with Clang, it becomes a
negative number. If `allocatedBytes_` is a small positive integer
when `incrementAllocatedBytes()` is called, `allocatedBytes_` will
underflow and result in an extremely large number. The assertion in
`incrementAllocatedBytes()` will fail.

This diff change the parameter type to uint32_t, and add
`decrementAllocatedBytes()` function to call when freeing memory.

Reviewed By: tmikov

Differential Revision: D89620141

fbshipit-source-id: 56a448ecda1a7de7160b0dfc439a12149759057c