Don't skip permanent skip list when "--test-skiplist" is passed
Summary:
With this change, we no longer skip the below three sections in
skiplist.json when `--test-skiplist` is passed:
- PERMANENT_SKIP_LIST
- UNSUPPORTED_FEATURES
- PERMANENT_UNSUPPORTED_FEATURES

The skiplist.json is updated in following diff.

Note that though we don't skip the "unsupported features" section, the
test runner won't add all tests with given feature into skiplist and
remove passed ones. It's better to remove a feature manually when it's
implemented.

Reviewed By: avp

Differential Revision: D89577066

fbshipit-source-id: aebe817011332b2616ee7e0ceca15a436946647e